maj next-auth v5
