+++ .prettierignore
@@ -1,0 +1,5 @@
+dist
+*.md
+package.json
+coverage
+.yalc
