+++ app/package.json
@@ -26,7 +26,7 @@
     "@hookform/resolvers": "^3.3.2",
     "@pedaki/api": "workspace:*",
     "@pedaki/auth": "workspace:*",
-    "@pedaki/design": "0.4.2",
+    "@pedaki/design": "0.4.3-beta.0",
     "@pedaki/logger": "workspace:*",
     "@t3-oss/env-nextjs": "^0.7.1",
     "@tanstack/react-query": "^4.36.1",
@@ -36,7 +36,7 @@
     "@trpc/server": "^10.44.1",
     "dayjs": "^1.11.10",
     "next": "14.0.4",
-    "next-auth": "^4.24.5",
+    "next-auth": "^5.0.0-beta.4",
     "next-international": "^1.1.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
