+++ app/src/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,2 @@
-import { authOptions } from '@pedaki/auth/server';
-import NextAuth from 'next-auth';
-
-const handler = NextAuth(authOptions) as Function;
-
-export { handler as GET, handler as POST };
+export { GET, POST } from '@pedaki/auth/server.ts';
+export const runtime = "nodejs" // default but we want to be explicit
