+++ packages/api/package.json
@@ -29,7 +29,7 @@
   "dependencies": {
     "@opentelemetry/api": "^1.7.0",
     "@pedaki/auth": "workspace:*",
-    "@pedaki/common": "0.4.2",
+    "@pedaki/common": "0.4.3-beta.0",
     "@pedaki/db": "workspace:*",
     "@pedaki/logger": "workspace:*",
     "@pedaki/services": "workspace:*",
