+++ packages/auth/package.json
@@ -28,11 +28,11 @@
   "license": "CC-BY-NC-SA-4.0",
   "dependencies": {
     "@auth/prisma-adapter": "^1.0.9",
-    "@pedaki/common": "0.4.2",
+    "@pedaki/common": "0.4.3-beta.0",
     "@pedaki/db": "workspace:*",
     "@pedaki/logger": "workspace:*",
     "@t3-oss/env-core": "^0.7.1",
-    "next-auth": "^4.24.5",
+    "next-auth": "5.0.0-beta.4",
     "superjson": "^2.2.1",
     "zod": "^3.22.4"
   },
