+++ packages/auth/src/edge.ts
@@ -1,0 +1,4 @@
+import NextAuth from "next-auth"
+import {baseAuthOptions} from "./index.ts";
+
+export const {  auth } = NextAuth(baseAuthOptions)
