+++ packages/services/package.json
@@ -20,7 +20,7 @@
   "author": "Nathan David <me@vahor.fr>",
   "license": "MIT",
   "dependencies": {
-    "@pedaki/common": "0.4.2",
+    "@pedaki/common": "0.4.3-beta.0",
     "@pedaki/db": "workspace:*",
     "@pedaki/logger": "workspace:*",
     "@pedaki/auth": "workspace:*",
